requirements/requirements.txt
Django==1.4.2
Pillow==1.7.8
South==0.7.6
Whoosh==2.4.1
argparse==1.2.1
distribute==0.6.24
django-appconf==0.5
django-compressor==1.2
django-debug-toolbar==0.9.4
django-filebrowser-no-grappelli==3.1.1
django-haystack==1.2.7
django-tinymce==1.5.1b2
ipdb==0.7
ipython==0.13.1
psycopg2==2.4.5
python-memcached==1.48
python-ptrace==0.6.4
pytz==2012h
sorl-thumbnail==11.12
wsgiref==0.1.2
